chore: Remove console.log statements and refactor code in authenticat…

+++ src/routes/postes.js
@@ -3,10 +3,11 @@ import { v2 as cloudinary } from "cloudinary";
 import fs from "fs-extra";
 import { db } from "../firebase.js";
 import { User } from "../auth/authentication.js";
+import helpers from "../auth/helpers.js";
 
 const postes = Router();
 
-postes.post("/add/pet", async (req, res) => {
+postes.post("/add/pet", helpers.isLoggedIn, async (req, res) => {
 
     const newPet = {
         name: req.body.name,
@@ -16,25 +17,38 @@ postes.post("/add/pet", async (req, res) => {
         weight: req.body.weight,
         idcollar: req.body.idcollar,
         address: req.body.address,
-        image: (await cloudinary.uploader.upload(req.file.path)).url
+        image: req.file != undefined ? (await cloudinary.uploader.upload(req.file.path)).url : null
     };
 
-    await fs.unlink(req.file.path);
+    if (req.file != undefined) {
+        await fs.unlink(req.file.path);
+    }
 
     let pet = await db.collection("pets").add(newPet);
 
-    await db.collection("users").doc(User).update({
-        pets: pet.id
+    let Users = await db.collection("users").get();
+
+    Users.forEach(async (doc) => {
+
+        if (doc.data().uid === User.uid) {
+            console.log("entrando");
+
+            let newPets = doc.data().pets;
+
+            newPets.push(pet.id);
+
+            await db.collection("users").doc(doc.id).update({ pets: newPets });
+        }
     });
-    
+
     res.redirect("/home");
 });
 
-postes.post("/edit/pet/:id", async (req, res) => {
+postes.post("/edit/pet/:id", helpers.isLoggedIn, async (req, res) => {
 
     const id = req.params.id;
 
-    const newPet = {
+    const newDataPet = {
         name: req.body.name,
         specie: req.body.specie,
         breed: req.body.breed,
@@ -42,29 +56,38 @@ postes.post("/edit/pet/:id", async (req, res) => {
         weight: req.body.weight,
         idcollar: req.body.idcollar,
         address: req.body.address,
-        image: (await cloudinary.uploader.upload(req.file.path)).url
+        image: req.file != undefined ? (await cloudinary.uploader.upload(req.file.path)).url : null
     };
 
-    await fs.unlink(req.file.path);
+    if (req.file != undefined) {
+        await fs.unlink(req.file.path);
+    }
 
-    await db.collection("pets").doc().update(newPet);
+    await db.collection("pets").doc(id).update(newDataPet);
 
     res.redirect("/home");
 });
 
-postes.post("/profile/edit", async (req, res) => {
+postes.post("/profile/edit", helpers.isLoggedIn, async (req, res) => {
 
     const newUserData = {
         name: req.body.name,
         phone: req.body.phone,
-        address: req.body.address,
-        email: req.body.email,
-        password: req.body.password,
+        address: req.body.address
     };
 
-    await db.collection("users").doc().update(newUserData);
+    let Users = await db.collection("users").get();
+
+    Users.forEach(async (doc) => {
+
+        if (doc.data().uid === User.uid) {
+            await db.collection("users").doc(doc.id).update(newUserData);
+        }
+
+    });
 
     res.redirect("/profile");
+
 });
 
 export default postes;
